feat[FE-119] Revamp and integrate API in wallet page
