+++ package.json
@@ -8,6 +8,7 @@
 		"@testing-library/jest-dom": "^5.16.5",
 		"@testing-library/react": "^13.4.0",
 		"@testing-library/user-event": "^13.5.0",
+		"atatus-js": "^4.5.0",
 		"axios": "^1.2.0",
 		"html-webpack-plugin": "^5.5.0",
 		"nanoid": "^4.0.0",
